pyproject.toml
[project]
name = "mapper-fivetran"
version = "0.0.1"
description = "Singer mapper Fivetran, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Reuben Frankel", email = "rfrankel@matatika.com" }]
keywords = [
    "ELT",
    "Mapper",
    "Fivetran",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license-files = [ "LICENSE" ]
requires-python = ">=3.9"
dependencies = [
    "pyhumps>=3.8.0",
    "singer-sdk~=0.48.0",
]

[project.optional-dependencies]
s3 = [
    "fs-s3fs~=1.1.1",
]

[project.scripts]
# CLI declaration
mapper-fivetran = 'mapper_fivetran.mapper:FivetranMapper.cli'

[dependency-groups]
dev = [
    "pytest>=8",
    "singer-sdk[testing]",
]

[tool.pytest.ini_options]
addopts = [
    "--durations=10",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true

[tool.ruff.lint]
ignore = [
    "ANN001",  # missing-type-function-argument
    "ANN002",  # missing-type-args
    "ANN003",  # missing-type-kwargs
    "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D1",
    "S101",
    "SLF001",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = [
    "hatchling>=1,<2",
]
build-backend = "hatchling.build"
